rename: verse_view widgets 폴더로 이동
